actually initialize kcrash properly
Summary:
to successfully use kcrash when linking with as-needed (which is a default
flag on many linux distros) one also needs to call KCrash::initialize.

without this kcrash may not be available for most users.

https://markmail.org/thread/zv5pheijaze72bzs

(kcrash is already in the link target list)

Test Plan: builds; links correctly

Reviewers: #plasma, ivan, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D17119