Remove pointless QDBusServiceWatcher
Summary:
KActivitymanagerd's ksmserver has a QDBusServiceWatcher in order to
recreate a QDBusInterface when the service goes away. There is no need
to do this, a call to a given method to a given service will work
regardless, dispatching to the correct client is all internal to
dbus-daemon.

QDBusAbstractInterface has it's own internal QDBusServiceWatcher anyway
so we don't get any behavioural differences if we check isValid.

This is worth fixing as this class is moved threads and we end up with
newly created children in a mess.

CCBUG: 413940

Test Plan:

Reviewers: #plasma, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25205