Implement Activities Runner as DBus Runner
This MR implements the activities runner using the DBus API.
The runner allows one to view and switch to other activities.

Because the data is already available in the kactivitymanagerd
process it makes very little sense to have a KRunner plugin which
fetches this data and produces the results when we can provide this
functionality in the actual running process.

See https://phabricator.kde.org/T12031 for more discussion
about the topic in general.