Avoid race condition on startup
The activities list needs to be loaded at the time the first DBus call
is processed. Otherwise kamd will return an empty list.

updateSortedActivityList was queued as it needs to run after some other
code in the main KAMD constructor, but this was racey as we could get
DBus traffic in the meantime.

This was relatively easy to hit as this is not in the main thread where
the connection is registered and there is a blocking DBus call in the
Activies constructor.

BUG: 422682


(cherry picked from commit 3cbd1b7b0fda5aa57411956c0cec0074163a4a8e)